reverse proxy for xqerl (#5)
 - start running reverse-proxy in the pod name podx
 - check site is reachable over port 443
 - or container action as a reverse proxy for xq container

+++ inc/content.mk
@@ -10,8 +10,7 @@ BuildTemplate := $(patsubst src/%.xq,build/%.tpl.txt,$(TemplateList))
 BuildDataMap  := $(patsubst src/%.json,build/%.map.txt,$(DataMapList))
 # build/html/%.html: build/data/%.cmark.txt
 PHONY: content
-content: $(BuildTemplate) $(BuildDataMap) $(BuildMarkdown)
-
+content: $(BuildMarkdown) $(BuildTemplate) $(BuildDataMap)
 PHONY: content-tar
 content-tar: deploy deploy/xqerl-database.tar
 
@@ -22,7 +21,7 @@ watch-content:
         inotifywait -qre close_write ./src/data/$(DOMAIN)/content/ || true; \
     done
 
-CONTENT := home/index
+CONTENT := home/articles/reverse-proxy-setup
 
 PHONY: content-view
 content-view:
@@ -53,27 +52,40 @@ content-list:
 
 build/data/%.cmark.txt: src/data/%.md
 	@[ -d $(dir $@) ] || mkdir -p $(dir $@)
+	@[ -d .tmp ] || mkdir -p .tmp
 	@echo '## $(notdir  $<) ##'
 	@bin/xq put $< | tee $@
-	@URI=http://localhost:8081/$(patsubst src/data/%.md,%,$<)
-	@echo "Page URL: $$URI"
+	@grep -q 'XDM item: document-node' $@
+
+xcxcxcxxc:
+ifndef GITHUB_ACTIONS
+	@POD_URI=http://localhost:8081/$(patsubst src/data/%.md,%,$<)
+	@echo "Internal Pod Page URL: $$POD_URI"
 	@#podman run --pod $(POD) --rm $(W3M) -dump_source -o accept_encoding='identity;q=0' $$URI
 	@$(DASH)
-	@podman run --pod $(POD) --rm $(W3M) -dump $$URI
+	@podman run --pod $(POD) --rm $(W3M) -dump $$POD_URI
 	@echo && $(DASH)
-	@#podman run --pod $(POD) --rm --mount $(MountAssets) $(ALPINE) mkdir -p $(patsubst src/data/%/$(notdir $<),%,$<)
-	@#podman run --pod $(POD) --rm --mount $(MountAssets) $(ALPINE) ls -alR .
-
+	@[ -e .tmp/example.com.pem ] \
+    || openssl s_client -showcerts -connect example.com:8443 </dev/null \
+		| sed -n -e '/-.BEGIN/,/-.END/ p' > .tmp/example.com.pem
+	@$(DASH)
+	@curl  -s -D - -o /dev/null --cacert .tmp/example.com.pem https://example.com:8443/$(patsubst src/data/example.com/content/%.md,%,$<)
+	@$(DASH)
+	@echo https://example.com:8443/$(patsubst src/data/example.com/content/%.md,%,$<)
+	@$(DASH)
+endif
 
 build/data/%.map.txt: src/data/%.json
 	@[ -d $(dir $@) ] || mkdir -p $(dir $@)
 	@echo '## $(notdir  $<) ##'
 	@bin/xq put $< | tee $@
+	@grep -q 'XDM item: map' $@
 
 build/data/%.tpl.txt: src/data/%.xq
 	@[ -d $(dir $@) ] || mkdir -p $(dir $@)
 	@echo '## $(notdir  $<) ##'
 	@bin/xq put $< | tee $@
+	@grep -q 'XDM item: function' $@
 
 deploy/xqerl-database.tar:
 	@[ -d $(dir $@) ] || mkdir -p $(dir $@)
